refactor: rename interflow package to fairness
Rename `pkg/epp/flowcontrol/framework/plugins/interflow` to
`pkg/epp/flowcontrol/framework/plugins/fairness`.

Update all references and package declarations. This aligns the
implementation package name with the `FairnessPolicy` interface.
