+++ pkg/epp/flowcontrol/framework/plugins/fairness/roundrobin.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package interflow
+package fairness
 
 import (
 	"context"
